Increase Org Scanning Interval
Increasing interval that concourse checks for new releases in github org
from `default: 1m` to `2m`.

This should help with rate limiting, decreasing the number of requests
to github's API.  Everytime a build is triggered concourse makes an auth
request in order to checkout the app. The increased usage of the
platform has lead to us hitting the limit.