Allow webhooks and reduce polling frequency on pipelines
This commit changes the default polling period for the github release resource
from 1 minute to 24 hours. The reason we can do this is the new webhook
dispatcher. The dispatcher will trigger a 'check' on this resource when we
receive a notification from a github org hook telling us there has been a new
release.

Existing pipelines have been updated in-place but this change is required so new
pipelines also get these changes.