update-pipelines-resource-types.py: handle no resource types
When running this on `alpha`, update of one of the pipelines failed.
Reason being this `aws-whoami` pipeline has a `resource_types` field
set to `null`/`None`.

The script was handling the case when this `resource_types` field
was not there but not when it was `null`.

I've tweaked the script to check if `resource_types` is truthy
before trying to manipulate its definition.